+++ src/main/java/com/example/nutriplan_backend/repository/PlanUsuarioRepository.java
@@ -10,6 +10,6 @@ import org.springframework.stereotype.Repository;
 @Repository
 
 // Interface para acceso a los métodos de JPA porque vamos a crear el CRUD
-public interface PlanUsuarioRepository extends JpaRepository<PlanUsuario, Long> {
-    List<PlanUsuario> findByUsuarioId(Long usuarioId);
+public interface PlanUsuarioRepository extends JpaRepository<PlanUsuario, Integer> {
+    List<PlanUsuario> findByUsuarioId(Integer usuarioId);
 }
